+++ examples/cli/example_zhmcrc.sh
@@ -1,0 +1,12 @@
+#!/bin/sh
+
+# In order to use the shell script examples in this folder,
+# you have to rename example_zhmcrc.sh to zhmcrc.sh and setup
+# your runtime configuration information like HMC_HOST, HMC_USER, etc.
+# When you source the file, environment variables are set
+# for your current shell.
+export HMC_HOST="Put here the HMC host"
+export HMC_USER="Put here HMC user"
+export CPC_NAME="Put here the name of the CPC"
+export PART_NAME="TESTPARTITION"
+